Render background pixels by x,y instead of by namatable index
Rather than rendering both nametables with scroll taken into account at
the beginning of vblank, render x,y from the cycle and scanlines during
clock ticks. This is still not quite correct, but does allow sprite zero
to be useful and makes split scrolling (horizontal only for now) to be
possible.

Sprite zero is still not right. For example the HUD in Super Mario Bros
now stays in place, but there are flickering pixels when rendering
nametable 2. The way to fix these issues is to further make rendering
PPU cycle accurate for both background and sprites.