Implement MMC1
Add support for 8x16 sprites. Turn cartridge into an interface and pull
the existing implementation out into a new nrom struct which implements
it. Add new mmc1 struct and implement it various prg and chr modes as
well as single page mirroring.

Tested implementation on the Legend of Zelda and Mega Man 2 and both
appear to be working well!