Stop using requestAnimationFrame
It's too hard to try and translate the screen refresh rate (raf tries
to synchronize with this) into 60 FPS. So instead of trying that
fire up a goroutine that infinitely awaits a time.Ticker. This appears
to work much better than any of the other things I've tried to get
close to a jitter-free rendering experience.